LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.measure.index-from-offset-edges.tentative-expected.txt

2d.text.measure.index-from-offset-edges.tentative

PASS Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction ltr and text align left.
FAIL Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction rtl and text align left. assert_equals: for -0123456789- multiple -2 expected 12 but got 0
PASS Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction ltr and text align center.
FAIL Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction rtl and text align center. assert_equals: for -0123456789- multiple -2 expected 12 but got 0
PASS Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction ltr and text align right.
FAIL Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction rtl and text align right. assert_equals: for -0123456789- multiple -2 expected 12 but got 0
PASS Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction ltr and text align start.
FAIL Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction rtl and text align start. assert_equals: for -0123456789- multiple -2 expected 12 but got 0
PASS Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction ltr and text align end.
FAIL Check that TextMetrics::getIndexFromOffset() gives correct edges when the requested point is outside the range, with direction rtl and text align end. assert_equals: for -0123456789- multiple -2 expected 12 but got 0
